+++ app/wyze_bridge.py
@@ -10,7 +10,7 @@ import threading
 import time
 import warnings
 from subprocess import PIPE, Popen
-from typing import List, NoReturn, Tuple, Union
+from typing import List, NoReturn, Optional, Tuple, Union
 
 import mintotp
 import paho.mqtt.publish
@@ -21,13 +21,13 @@ import wyzecam
 
 class WyzeBridge:
     def __init__(self) -> None:
-        print("🚀 STARTING DOCKER-WYZE-BRIDGE v1.4.1\n")
+        print("🚀 STARTING DOCKER-WYZE-BRIDGE v1.4.2\n")
         signal.signal(signal.SIGTERM, lambda n, f: self.clean_up())
         self.hass: bool = bool(os.getenv("HASS"))
         self.on_demand: bool = bool(os.getenv("ON_DEMAND"))
-        self.timeout: int = int(env_bool("RTSP_READTIMEOUT", 15).replace("s", ""))
-        self.connect_timeout: int = int(env_bool("CONNECT_TIMEOUT", 20))
-        self.keep_bad_frames: bool = env_bool("KEEP_BAD_FRAMES", False)
+        self.timeout: int = env_bool("RTSP_READTIMEOUT", 15, style="int")
+        self.connect_timeout: int = env_bool("CONNECT_TIMEOUT", 20, style="int")
+        self.keep_bad_frames: bool = env_bool("KEEP_BAD_FRAMES", style="bool")
         self.healthcheck: bool = bool(os.getenv("HEALTHCHECK"))
         self.token_path: str = "/config/wyze-bridge/" if self.hass else "/tokens/"
         self.img_path: str = "/%s/" % env_bool("IMG_DIR", "img").strip("/")
@@ -66,7 +66,7 @@ class WyzeBridge:
         """Start all streams and keep them alive."""
         for cam_name in self.streams:
             self.start_stream(cam_name)
-        cooldown = int(env_bool("OFFLINE_TIME", 10))
+        cooldown = env_bool("OFFLINE_TIME", 10, style="int")
         while self.streams and not self.stop_flag.is_set():
             refresh_cams = True
             for name, stream in list(self.streams.items()):
@@ -108,6 +108,7 @@ class WyzeBridge:
             if hasattr(proc, "alive") and proc.alive():
                 proc.terminate()
                 proc.join()
+        offline = bool(self.streams[name].get("sleep"))
         cam = next(c for c in self.cameras if c.nickname == name)
         model = model_names.get(cam.product_model, cam.product_model)
         log.info(f"🎉 Connecting to WyzeCam {model} - {name} on {cam.ip} (1/3)")
@@ -115,7 +116,7 @@ class WyzeBridge:
 
         stream = multiprocessing.Process(
             target=self.start_tutk_stream,
-            args=(cam, self.stop_flag, connected),
+            args=(cam, self.stop_flag, connected, offline),
             name=name,
         )
         self.streams[name] = {
@@ -292,47 +293,6 @@ class WyzeBridge:
         if pas := env_bool(path + "READPASS", os.getenv("RTSP_PATHS_ALL_READPASS")):
             os.environ[path + "READPASS"] = pas
 
-    def mqtt_discovery(self, cam) -> None:
-        """Add cameras to MQTT if enabled."""
-        if not env_bool("MQTT_HOST"):
-            return
-        uri = f"{clean_name(cam.nickname)}"
-        msgs = [(f"wyzebridge/{uri}/state", "offline")]
-        if env_bool("MQTT_DTOPIC"):
-            topic = f"{os.getenv('MQTT_DTOPIC')}/camera/{cam.mac}/config"
-            payload = {
-                "uniq_id": "WYZE" + cam.mac,
-                "name": "Wyze Cam " + cam.nickname,
-                "topic": f"wyzebridge/{uri}/image",
-                "json_attributes_topic": f"wyzebridge/{uri}/attributes",
-                "availability_topic": f"wyzebridge/{uri}/state",
-                "icon": "mdi:image",
-                "device": {
-                    "connections": [["mac", cam.mac]],
-                    "identifiers": cam.mac,
-                    "manufacturer": "Wyze",
-                    "model": cam.product_model,
-                    "sw_version": cam.firmware_ver,
-                    "via_device": "docker-wyze-bridge",
-                },
-            }
-            msgs.append((topic, json.dumps(payload)))
-        mqauth = os.getenv("MQTT_AUTH", ":").split(":")
-        mqhost = os.getenv("MQTT_HOST", "localhost").split(":")
-        try:
-            paho.mqtt.publish.multiple(
-                msgs,
-                hostname=mqhost[0],
-                port=int(mqhost[1]) if len(mqhost) > 1 else 1883,
-                auth=(
-                    {"username": mqauth[0], "password": mqauth[1]}
-                    if env_bool("MQTT_AUTH")
-                    else None
-                ),
-            )
-        except Exception as ex:
-            log.warning(f"[MQTT] {ex}")
-
     def get_filtered_cams(self) -> None:
         """Get all cameras that are enabled."""
         cams = self.get_wyze_data("cameras")
@@ -361,7 +321,7 @@ class WyzeBridge:
         print(f"\n🎬 STARTING {msg} {total} CAMERAS")
         for cam in cams:
             self.add_rtsp_path(cam)
-            self.mqtt_discovery(cam)
+            mqtt_discovery(cam)
             self.save_api_thumb(cam)
             self.streams[cam.nickname] = {}
 
@@ -376,28 +336,29 @@ class WyzeBridge:
             log.info("starting rtsp-simple-server")
         self.rtsp = Popen(["/app/rtsp-simple-server", "/app/rtsp-simple-server.yml"])
 
-    def start_tutk_stream(self, cam: wyzecam.WyzeCamera, stop_flag, connected) -> None:
+    def start_tutk_stream(
+        self, cam: wyzecam.WyzeCamera, stop_flag, connected, offline
+    ) -> None:
         """Connect and communicate with the camera using TUTK."""
         uri = clean_name(cam.nickname, upper=True)
         exit_code = 1
-        audio = env_bool(f"ENABLE_AUDIO_{uri}", env_bool("ENABLE_AUDIO", False))
+        audio = env_bool(f"ENABLE_AUDIO_{uri}", env_bool("ENABLE_AUDIO"), style="bool")
         try:
             with wyzecam.WyzeIOTC() as wyze_iotc, wyzecam.WyzeIOTCSession(
                 wyze_iotc.tutk_platform_lib,
                 self.user,
                 cam,
                 *(get_env_quality(uri, cam.product_model)),
-                enable_audio=bool(audio),
+                enable_audio=audio,
                 connect_timeout=self.connect_timeout,
             ) as sess:
                 connected.set()
-                a_codec = sess.get_audio_codec() if audio else None
-                fps = check_cam_sess(sess, uri, a_codec)
+                fps, audio = get_cam_params(sess, uri, audio)
                 audio_thread = threading.Thread(
                     target=sess.recv_audio_frames, args=(uri, fps), name=uri + "_AUDIO"
                 )
                 with Popen(
-                    get_ffmpeg_cmd(uri, cam.product_model, audio, a_codec), stdin=PIPE
+                    get_ffmpeg_cmd(uri, cam.product_model, audio), stdin=PIPE
                 ) as ffmpeg:
                     if audio:
                         audio_thread.start()
@@ -405,15 +366,20 @@ class WyzeBridge:
                         stop_flag, self.keep_bad_frames, self.timeout, fps
                     ):
                         ffmpeg.stdin.write(frame)
-        except Exception as ex:
+        except wyzecam.TutkError as ex:
             log.warning(ex)
-            if ex.args[0] == -13:  # IOTC_ER_TIMEOUT
+            set_cam_offline(uri, ex, offline)
+            if ex.code == -13:  # IOTC_ER_TIMEOUT
                 time.sleep(2)
-            elif ex.args[0] in (-19, -68, -90):
-                exit_code = abs(ex.args[0])
-            elif ex.args[0] == "Authentication did not succeed! {'connectionRes': '2'}":
+            elif ex.code in (-19, -68, -90):
+                exit_code = abs(ex.code)
+        except ValueError as ex:
+            log.warning(ex)
+            if ex.args[0] == "ENR_AUTH_FAILED":
                 log.warning("⏰ Expired ENR?")
                 exit_code = 19
+        except Exception as ex:
+            log.warning(ex)
         else:
             log.warning("Stream is down.")
         finally:
@@ -455,10 +421,21 @@ model_names = {
 }
 
 
-def env_bool(env: str, false: str = "") -> str:
+def env_bool(env: str, false="", true="", style="") -> str:
     """Return env variable or empty string if the variable contains 'false' or is empty."""
-    env = os.getenv(env.upper().replace("-", "_"), "").lower().replace("false", "")
-    return env or false
+    env_value = os.getenv(env.upper().replace("-", "_"), "")
+    value = env_value.lower().replace("false", "").strip("'\" \n\t\r")
+    if style.lower() == "bool":
+        return bool(value or false)
+    if style.lower() == "int":
+        return int("".join(filter(str.isdigit, value or str(false))) or 0)
+    if style.lower() == "upper" and value:
+        return value.upper()
+    if style.lower() == "original" and value:
+        return os.getenv(env.upper().replace("-", "_"))
+    if true and value:
+        return true
+    return value or false
 
 
 def env_list(env: str) -> list:
@@ -481,11 +458,7 @@ def env_filter(cam) -> bool:
 
 def get_env_quality(uri: str, cam_model: str) -> Tuple[int, int]:
     """Get preferred resolution and bitrate from env."""
-    env_quality = (
-        env_bool(f"QUALITY_{uri}", env_bool("QUALITY", "na"))
-        .strip("'\"\n ")
-        .ljust(3, "0")
-    )
+    env_quality = env_bool(f"QUALITY_{uri}", env_bool("QUALITY", "na")).ljust(3, "0")
     env_bit = int(env_quality[2:])
     frame_size = 1 if env_quality[:2] == "sd" else 0
     if doorbell := (cam_model == "WYZEDB3"):
@@ -522,11 +495,17 @@ def check_net_mode(session_mode: int, uri: str) -> str:
     return mode
 
 
-def check_cam_sess(sess: wyzecam.WyzeIOTCSession, uri: str, audio: tuple = None) -> int:
-    """Check cam session and return connection mode, firmware, and wifidb from camera."""
+def get_cam_params(
+    sess: wyzecam.WyzeIOTCSession, uri: str, audio: bool
+) -> Tuple[int, Optional[dict]]:
+    """Check session and return fps and audio codec from camera."""
     mode = check_net_mode(sess.session_check().mode, uri)
+    if env_bool("IOTC_TCP"):
+        sess.tutk_platform_lib.IOTC_TCPRelayOnly_TurnOn()
+        print(sess.session_check())
     frame_size = "SD" if sess.preferred_frame_size == 1 else "HD"
     bit_frame = f"{sess.preferred_bitrate}kb/s {frame_size} stream"
+    fps = 20
     if video_param := sess.camera.camera_info.get("videoParm", False):
         if fps := int(video_param.get("fps", 0)):
             if fps % 5 != 0:
@@ -544,17 +523,19 @@ def check_cam_sess(sess: wyzecam.WyzeIOTCSession, uri: str, audio: tuple = None)
     wifi = sess.camera.camera_info["basicInfo"].get("wifidb", "NA")
     if "netInfo" in sess.camera.camera_info:
         wifi = sess.camera.camera_info["netInfo"].get("signal", wifi)
-    # return mode, firmware, wifi
+    if audio:
+        codec, rate = sess.get_audio_codec()
+        codec_str = codec.replace("s16le", "PCM")
+        if codec_out := env_bool("AUDIO_CODEC", "AAC" if "s16le" in codec else ""):
+            codec_str += " > " + codec_out
+        audio: dict = {"codec": codec, "rate": rate, "codec_out": codec_out.lower()}
     log.info(f"📡 Getting {bit_frame} via {mode} (WiFi: {wifi}%) FW: {firmware} (2/3)")
     if audio:
-        a_codec = audio[0].replace("s16le", "pcm > aac").upper()
-        log.info(f"🔊 Audio Enabled - {a_codec}/{audio[1]:,}Hz")
-    return fps or 20
+        log.info(f"🔊 Audio Enabled - {codec_str.upper()}/{rate:,}Hz")
+    return fps, audio
 
 
-def get_ffmpeg_cmd(
-    uri: str, cam_model: str, audio: bool = False, a_codec: tuple = None
-) -> list:
+def get_ffmpeg_cmd(uri: str, cam_model: str, audio: Optional[dict]) -> list:
     """Return the ffmpeg cmd with options from the env."""
     lib264 = (
         ["libx264", "-filter:v", "transpose=1", "-b:v", "3000K"]
@@ -562,22 +543,21 @@ def get_ffmpeg_cmd(
         + ["-force_key_frames", "expr:gte(t,n_forced*2)"]
     )
     flags = "-fflags +genpts+flush_packets+nobuffer -flags low_delay"
-    rotate = cam_model == "WYZEDB3" and env_bool("ROTATE_DOOR", False)
+    rotate = cam_model == "WYZEDB3" and env_bool("ROTATE_DOOR")
     livestream = get_livestream_cmd(uri)
     audio_in = "-f lavfi -i anullsrc=cl=mono" if livestream else ""
     audio_out = "aac"
-    if audio and a_codec:
-        audio_in = f"-f {a_codec[0]} -ar {a_codec[1]} -i /tmp/{uri.lower()}.wav"
-        audio_out = env_bool("AUDIO_CODEC", "copy")
-        audio_out = "aac" if audio_out == "copy" and "s16le" in a_codec else audio_out
+    if audio and "codec" in audio:
+        audio_in = f"-f {audio['codec']} -ar {audio['rate']} -i /tmp/{uri.lower()}.wav"
+        audio_out = audio["codec_out"] or "copy"
         a_filter = ["-filter:a"] + env_bool("AUDIO_FILTER", "volume=5").split()
     av_select = "select=" + ("v,a" if audio else "v")
-    rtsp_proto = "udp" if "udp" in env_bool("RTSP_PROTOCOLS").lower() else "tcp"
+    rtsp_proto = "udp" if "udp" in env_bool("RTSP_PROTOCOLS") else "tcp"
     rtsp_ss = f"[{av_select}:f=rtsp:rtsp_transport={rtsp_proto}]rtsp://0.0.0.0:8554/{uri.lower()}"
 
-    cmd = env_bool(f"FFMPEG_CMD_{uri}", env_bool("FFMPEG_CMD", "")).strip(
-        "'\"\n "
-    ).format(cam_name=uri.lower(), CAM_NAME=uri, audio_in=audio_in).split() or (
+    cmd = env_bool(f"FFMPEG_CMD_{uri}", env_bool("FFMPEG_CMD")).format(
+        cam_name=uri.lower(), CAM_NAME=uri, audio_in=audio_in
+    ).split() or (
         ["-loglevel", "verbose" if env_bool("DEBUG_FFMPEG") else "error"]
         + env_bool(f"FFMPEG_FLAGS_{uri}", env_bool("FFMPEG_FLAGS", flags))
         .strip("'\"\n ")
@@ -603,7 +583,7 @@ def get_ffmpeg_cmd(
 
 def get_record_cmd(uri: str, av_select: str) -> str:
     """Check if recording is enabled and return ffmpeg tee cmd."""
-    if not env_bool(f"RECORD_{uri}", env_bool("RECORD_ALL", False)):
+    if not env_bool(f"RECORD_{uri}", env_bool("RECORD_ALL")):
         return ""
     seg_time = env_bool("RECORD_LENGTH", "60")
     file_name = "{CAM_NAME}_%Y-%m-%d_%H-%M-%S_%Z"
@@ -627,16 +607,96 @@ def get_record_cmd(uri: str, av_select: str) -> str:
 def get_livestream_cmd(uri: str) -> str:
     """Check if livestream is enabled and return ffmpeg tee cmd."""
     cmd = ""
-    if len(yt_key := env_bool(f"YOUTUBE_{uri}")) > 5:
+    if len(key := env_bool(f"YOUTUBE_{uri}", style="original")) > 5:
         log.info("📺 YouTube livestream enabled")
-        cmd += f"|[f=flv:select=v,a]rtmp://a.rtmp.youtube.com/live2/{yt_key}"
-    if len(fb_key := env_bool(f"FACEBOOK_{uri}")) > 5:
+        cmd += f"|[f=flv:select=v,a]rtmp://a.rtmp.youtube.com/live2/{key}"
+    if len(key := env_bool(f"FACEBOOK_{uri}", style="original")) > 5:
         log.info("📺 Facebook livestream enabled")
-        cmd += f"|[f=flv:select=v,a]rtmps://live-api-s.facebook.com:443/rtmp/{fb_key}"
-    if len(tee_cmd := env_bool(f"LIVESTREAM_{uri}")) > 5:
-        log.info(f"📺 Custom ({tee_cmd}) livestream enabled")
-        cmd += f"|[f=flv:select=v,a]{tee_cmd}"
+        cmd += f"|[f=flv:select=v,a]rtmps://live-api-s.facebook.com:443/rtmp/{key}"
+    if len(key := env_bool(f"LIVESTREAM_{uri}", style="original")) > 5:
+        log.info(f"📺 Custom ({key}) livestream enabled")
+        cmd += f"|[f=flv:select=v,a]{key}"
     return cmd
+
+
+def set_cam_offline(uri: str, error: wyzecam.TutkError, offline: bool) -> None:
+    """Do something when camera goes offline."""
+
+    state = "offline" if error.code == -90 else error.name
+
+    mqtt_status = [(f"wyzebridge/{uri.lower()}/state", state)]
+    if error.code == -90:
+        mqtt_status.append((f"wyzebridge/{uri.lower()}/offline", True))
+
+    send_mqtt(mqtt_status)
+
+    if str(error.code) not in env_bool("OFFLINE_ERRNO", "-90"):
+        return
+    if offline:  # Don't resend if previous state was offline.
+        return
+
+    if ":" in (ifttt := env_bool("OFFLINE_IFTTT", style="original")):
+        event, key = ifttt.split(":")
+        url = f"https://maker.ifttt.com/trigger/{event}/with/key/{key}"
+        data = {"value1": uri, "value2": error.code, "value3": error.name}
+        try:
+            resp = requests.post(url, data)
+            resp.raise_for_status()
+        except requests.exceptions.HTTPError as ex:
+            log.warning(f"[IFTTT] {ex}")
+        else:
+            log.info(f"[IFTTT] 📲 Sent webhook trigger to {event}")
+
+
+def mqtt_discovery(cam) -> None:
+    """Add cameras to MQTT if enabled."""
+    if not env_bool("MQTT_HOST"):
+        return
+    msgs = [
+        (f"wyzebridge/{clean_name(cam.nickname)}/state", "disconnected"),
+        (f"wyzebridge/{clean_name(cam.nickname)}/offline", None, 0, True),
+    ]
+    if env_bool("MQTT_DTOPIC"):
+        topic = f"{os.getenv('MQTT_DTOPIC')}/camera/{cam.mac}/config"
+        payload = {
+            "uniq_id": "WYZE" + cam.mac,
+            "name": "Wyze Cam " + cam.nickname,
+            "topic": f"wyzebridge/{uri}/image",
+            "json_attributes_topic": f"wyzebridge/{uri}/attributes",
+            "availability_topic": f"wyzebridge/{uri}/state",
+            "icon": "mdi:image",
+            "device": {
+                "connections": [["mac", cam.mac]],
+                "identifiers": cam.mac,
+                "manufacturer": "Wyze",
+                "model": cam.product_model,
+                "sw_version": cam.firmware_ver,
+                "via_device": "docker-wyze-bridge",
+            },
+        }
+        msgs.append((topic, json.dumps(payload)))
+    send_mqtt(msgs)
+
+
+def send_mqtt(messages: list) -> None:
+    """Publish a message to the MQTT server."""
+    if not env_bool("MQTT_HOST"):
+        return
+    m_auth = os.getenv("MQTT_AUTH", ":").split(":")
+    m_host = os.getenv("MQTT_HOST", "localhost").split(":")
+    try:
+        paho.mqtt.publish.multiple(
+            messages,
+            hostname=m_host[0],
+            port=int(m_host[1]) if len(m_host) > 1 else 1883,
+            auth=(
+                {"username": m_auth[0], "password": m_auth[1]}
+                if env_bool("MQTT_AUTH")
+                else None
+            ),
+        )
+    except Exception as ex:
+        log.warning(f"[MQTT] {ex}")
 
 
 def setup_hass():
@@ -681,7 +741,7 @@ if __name__ == "__main__":
         logging.getLogger().setLevel(debug_level)
     log = logging.getLogger("WyzeBridge")
     log.setLevel(debug_level if "DEBUG_LEVEL" in os.environ else logging.INFO)
-    if env_bool("DEBUG_FRAMES"):
+    if env_bool("DEBUG_FRAMES") or env_bool("DEBUG_LEVEL"):
         warnings.simplefilter("always")
     warnings.formatwarning = lambda msg, *args, **kwargs: f"WARNING: {msg}"
     logging.captureWarnings(True)