+++ app/wyzecam/tutk/tutk.py
@@ -288,9 +288,10 @@ class TutkError(RuntimeError):
         -31000: "TUNNEL_ER_UNDEFINED",
     }
 
-    def __init__(self, code):
+    def __init__(self, code, data=None):
         super().__init__(code)
         self.code = code
+        self.data = data
 
     @property
     def name(self):
@@ -375,8 +376,14 @@ class SInfoStructEx(FormattedStructure):
             "is_secure",
             c_uint8,
         ),  # 0: The IOTC session is in non-secure mode, 1: The IOTC session is in secure mode
-        ("local_nat_type", c_uint8),  # The local NAT type.
-        ("remote_nat_type", c_uint8),  # The remote NAT type.
+        (
+            "local_nat_type",
+            c_uint8,
+        ),  # The local NAT type, 0: Unknown type, 1: Type 1, 2: Type 2, 3: Type 3, 10: TCP only
+        (
+            "remote_nat_type",
+            c_uint8,
+        ),  # The remote NAT type, 0: Unknown type, 1: Type 1, 2: Type 2, 3: Type 3, 10: TCP only
         ("relay_type", c_uint8),  # 0: Not Relay, 1: UDP Relay, 2: TCP Relay
         (
             "net_state",
@@ -905,8 +912,7 @@ def iotc_check_device_online(
         c_uint(timeout_ms),
         c_int32(),
     )
-    # return status, device_out
-    return status
+    return status, device_out
 
 
 def iotc_connect_by_uid_parallel(